ref: use module import * as FormRunner to shorten type names
