+++ src/server/components/formRunner/lawyers.ts
@@ -1,11 +1,16 @@
 import { LawyerListItemGetObject, ServiceType } from "server/models/types";
-import {
-  FormRunnerField,
-  FormRunnerQuestion,
-} from "server/components/formRunner/types";
+import { Field, Question } from "server/components/formRunner/types";
 import { parseJsonFormData } from "server/components/formRunner/helpers";
 import { get } from "lodash";
 
+// TODO:- use type mapping or keyof operator instead of simple object type
+/**
+ * Maps {@link LawyersFormWebhookData} properties to a ListItem.
+ * @example
+ * To get the organisationName from a `ListItem`, `get(listItem, formRunnerFields.organisationName)`
+ * is equivalent to `get(listItem, "jsonData.organisationName")`
+ * is equivalent to `listItem.jsonData.organisationName`.
+ */
 const FormRunnerFields: { [key: string]: string } = {
   speakEnglish: "jsonData.speakEnglish",
   contactName: "jsonData.contactName",
@@ -32,13 +37,14 @@ const FormRunnerFields: { [key: string]: string } = {
   publicEmailAddress: "jsonData.publicEmailAddress",
 };
 
+// TODO: use the SerialisedWebhookData as the return type
 export async function generateFormRunnerWebhookData(
   listItem: LawyerListItemGetObject,
   isUnderTest?: boolean
-): Promise<Array<Partial<FormRunnerQuestion>>> {
+): Promise<Array<Partial<Question>>> {
   const questions = await parseJsonFormData(ServiceType.lawyers, isUnderTest);
   questions.forEach((question) => {
-    question.fields?.forEach((field: FormRunnerField) => {
+    question.fields?.forEach((field: Field) => {
       field.answer = get(listItem, FormRunnerFields[field.key]);
     });
   });
