+++ src/server/models/listItem/providers/deserialisers/Lawyer.deserialiser.ts
@@ -1,0 +1,11 @@
+import { LawyersFormWebhookData, WebhookDeserialiser } from "./types";
+
+export const lawyerDeserialiser: WebhookDeserialiser<LawyersFormWebhookData> = (
+  webhookData
+) => {
+  const { areasOfLaw = [], ...rest } = webhookData;
+  return {
+    areasOfLaw: areasOfLaw.filter(Boolean),
+    ...rest,
+  };
+};
